Make the L2 associativity twice as big as the L1's
On a benchmark that should exactly fit in the hierarchy, I had a L2 miss
ratio bigger than it was supposed to be. This happened when L2 was twice
as big as L1 but had the same associativity.

In this case, L2 has exactly twice as many sets as L1, which means that
L2 needs one more bit than L1 to index the cache. A consequence of this
is that a cache line in L1 can be mapped in either of two sets in L2.
On a L2 hit, the cache line is invalidated in L2, and brought to L1,
while a victim from L1 is installed in L2. If this victim is not
installed in the set where a line was just invalidated, it will evict
useful data

Thanks to my colleagues Vasilis and Andreas for spotting this!